feat: copy commit feature in a "fancier" way
